[feat] 스터디 정보 API 개발

+++ src/main/java/whiteship/toyproject/livestudy/common/repository/StudyCommentSiteRepository.java
@@ -1,9 +1,10 @@
 package whiteship.toyproject.livestudy.common.repository;
 
+import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import whiteship.toyproject.livestudy.common.model.StudyComment;
 import whiteship.toyproject.livestudy.common.model.StudyCommentSite;
 
-public interface StudyCommentSiteRepository extends JpaRepository<StudyCommentSite, Integer> {
+public interface StudyCommentSiteRepository extends JpaRepository<StudyCommentSite, Long> {
 
 }
